refactored restart schedule tests
